Fix audio distortion and align YM2612 logic with reference
Refactored mixing architecture to use integer buffers in `Ym2612Chip` and `PsgChip` to prevent overflow and wraparound distortion during channel summation.
Updated `VirtualSynthesizer` to handle mixing in high precision before final clamping.
Aligned YM2612 SSG-EG LFO application order and Feedback logic with `ym2612.c` reference for accurate sound reproduction.
Implemented DAC Highpass filter and removed non-standard `attackRamp` logic.
Fixed `computeModulationInput` and `computeCarrierSum` helpers to match documentation, ensuring algorithm routing tests pass.
Updated all relevant unit tests to support integer buffers and new output levels.

+++ src/test/java/uk/co/jamesj999/sonic/tests/TestYm2612TimerCSM.java
@@ -43,10 +43,10 @@ public class TestYm2612TimerCSM {
         }
 
         // Verify it is silent initially (no Key On)
-        short[] buffer = new short[100];
+        int[] buffer = new int[100];
         chip.render(buffer);
         boolean silent = true;
-        for (short s : buffer) {
+        for (int s : buffer) {
             if (s != 0) {
                 silent = false;
                 break;
@@ -72,11 +72,11 @@ public class TestYm2612TimerCSM {
         // Samples to overflow = 4190208 / 4948 ~ 846 samples.
 
         // Let's render 2000 samples to be sure.
-        buffer = new short[2000];
+        buffer = new int[2000];
         chip.render(buffer);
 
         boolean hasSound = false;
-        for (short s : buffer) {
+        for (int s : buffer) {
             if (s != 0) {
                 hasSound = true;
                 break;
